requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
pyzmq==25.0.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
Werkzeug==2.2.3
zmq==0.0.0
